Move helpers to src/app/lib/utilities
