+++ src/app/containers/Article/index.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shape } from 'prop-types';
-import compose from '../../helpers/compose';
+import compose from '../../lib/utilities/compose';
 import articlePropTypes from '../../models/propTypes/article';
 import ArticleMain from '../ArticleMain';
