+++ src/app/containers/PageViewAnalytics/labelHelpers/index.test.js
@@ -1,10 +1,10 @@
 import Cookie from 'js-cookie';
-import onClient from '../../../helpers/onClient';
+import onClient from '../../../lib/utilities/onClient';
 import { setWindowValue, resetWindowValue } from '../../../../testHelpers';
 
 let isOnClient = true;
 
-jest.mock('../../../helpers/onClient', () => jest.fn());
+jest.mock('../../../lib/utilities/onClient', () => jest.fn());
 onClient.mockImplementation(() => isOnClient);
 
 const {
