--- src/app/helpers/README.md
@@ -1,3 +1,0 @@
-# Helpers/utils
-
-Currently we do not differentiate between helpers and utils, put them all here. The only exception is models which have their own directory.
